[add_zenpen_style] Structure in different files by component.
